+++ hipamd/src/hip_table_interface.cpp
@@ -2063,3 +2063,8 @@ hipError_t hipKernelGetLibrary(hipLibrary_t* library, hipKernel_t kernel) {
 hipError_t hipKernelGetName(const char** name, hipKernel_t kernel) {
   return hip::GetHipDispatchTable()->hipKernelGetName_fn(name, kernel);
 }
+hipError_t hipKernelGetParamInfo(hipKernel_t kernel, size_t paramIndex, size_t* paramOffset,
+                                 size_t* paramSize) {
+  return hip::GetHipDispatchTable()->hipKernelGetParamInfo_fn(kernel, paramIndex, paramOffset,
+                                                              paramSize);
+}